+++ tests/testthat/test-integer64.R
@@ -377,14 +377,14 @@ test_that("all.equal works", {
 test_that("allNA and anyNA", {
 
   # Hopefully allNA is added to the base package like anyNA is.
-  expect_warning(expect_identical(allNA(c(1L, 1L)), FALSE), "Please promote that `allNA[(])`")
-  expect_warning(expect_identical(allNA(c(1L, NA)), FALSE), "Please promote that `allNA[(])`")
-  expect_warning(expect_identical(allNA(c(NA, NA)), TRUE), "Please promote that `allNA[(])`")
-  expect_warning(expect_identical(allNA(integer()), FALSE), "Please promote that `allNA[(])`")
-  expect_no_warning(expect_identical(allNA(as.integer64(c(1L, 1L))), FALSE))
-  expect_no_warning(expect_identical(allNA(as.integer64(c(1L, NA))), FALSE))
-  expect_no_warning(expect_identical(allNA(as.integer64(c(NA, NA))), TRUE))
-  expect_no_warning(expect_identical(allNA(integer64()), FALSE))
+  expect_identical(bit64.dev:::allNA(c(1L, 1L)), FALSE)
+  expect_identical(bit64.dev:::allNA(c(1L, NA)), FALSE)
+  expect_identical(bit64.dev:::allNA(c(NA, NA)), TRUE)
+  expect_identical(bit64.dev:::allNA(integer()), FALSE)
+  expect_no_warning(expect_identical(bit64.dev:::allNA(as.integer64(c(1L, 1L))), FALSE))
+  expect_no_warning(expect_identical(bit64.dev:::allNA(as.integer64(c(1L, NA))), FALSE))
+  expect_no_warning(expect_identical(bit64.dev:::allNA(as.integer64(c(NA, NA))), TRUE))
+  expect_no_warning(expect_identical(bit64.dev:::allNA(integer64()), FALSE))
 
   expect_identical(anyNA(as.integer64(c(1L, 1L))), anyNA(c(1L, 1L)))
   expect_identical(anyNA(as.integer64(c(1L, NA))), anyNA(c(1L, NA)))