+++ NEWS.md
@@ -46,7 +46,7 @@
 ## NOTES
 
 1. {bit64} no longer prints any start-up messages through an `.onAttach()` hook (#106). Thanks @hadley for the request.
-2. `anyNA` is supported for `integer64` and `allNA` is added.
+2. `anyNA` is supported for `integer64`.
 
 # bit64 4.6.0-1 (2025-01-16)